refactor: Lotto의 reward를 LottoWinningNumbers의 reward로 변경
