+++ src/test/java/com/lotto/LottoWinningNumbersTest.java
@@ -1,16 +1,23 @@
 package com.lotto;
 
+import com.lotto.domain.Lotto;
 import com.lotto.domain.LottoNumber;
+import com.lotto.domain.LottoReward;
 import com.lotto.domain.LottoWinningNumbers;
 import com.lotto.exception.IllegalInputException;
 import com.lotto.exception.IllegalLottoCountException;
 import com.lotto.exception.LottoNumberOutOfBoundsException;
 import org.assertj.core.api.Assertions;
+import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 
-import static org.assertj.core.api.Assertions.*;
+import java.util.Arrays;
+import java.util.TreeSet;
+
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 public class LottoWinningNumbersTest {
     @Test
@@ -50,5 +57,39 @@ public class LottoWinningNumbersTest {
     void createLottoWinningNumbers_잘못된입력데이터_음수_예외(String input) {
         assertThatThrownBy(() -> LottoWinningNumbers.createLottoWinningNumbers(input))
                 .isInstanceOf(LottoNumberOutOfBoundsException.class);
+    }
+
+    @DisplayName("로또 보상금 확인 1등")
+    @Test
+    void reward_성공() {
+        LottoWinningNumbers lottoWinningNumbers = LottoWinningNumbers.createLottoWinningNumbers("1,2,3,4,5,6");
+        LottoNumber number1 = new LottoNumber(1);
+        LottoNumber number2 = new LottoNumber(2);
+        LottoNumber number3 = new LottoNumber(3);
+        LottoNumber number4 = new LottoNumber(4);
+        LottoNumber number5 = new LottoNumber(5);
+        LottoNumber number6 = new LottoNumber(6);
+        Lotto lotto = new Lotto(new TreeSet<>(Arrays.asList(number1, number2, number3, number4, number5, number6)));
+
+        LottoReward reward = lottoWinningNumbers.reward(lotto);
+
+        assertThat(reward).isEqualTo(LottoReward.SIX);
+    }
+
+    @DisplayName("로또 보상금 확인 담청실패")
+    @Test
+    void reward_lose() {
+        LottoNumber number1 = new LottoNumber(1);
+        LottoNumber number2 = new LottoNumber(2);
+        LottoNumber number3 = new LottoNumber(3);
+        LottoNumber number4 = new LottoNumber(4);
+        LottoNumber number5 = new LottoNumber(5);
+        LottoNumber number6 = new LottoNumber(6);
+        Lotto lotto = new Lotto(new TreeSet<>(Arrays.asList(number1, number2, number3, number4, number5, number6)));
+
+        LottoWinningNumbers lottoWinningNumbers = LottoWinningNumbers.createLottoWinningNumbers("7,8,9,10,11,12");
+        LottoReward reward = lottoWinningNumbers.reward(lotto);
+
+        Assertions.assertThat(reward).isEqualTo(LottoReward.NOTHING);
     }
 }
